+++ Gitagram/View/LoginView/LoginView.swift
@@ -8,10 +8,7 @@
 import SwiftUI
 import FirebaseAuth
 struct LoginView: View {
-    var provider = OAuthProvider(providerID: "github.com")
-    @State private var oauthProvider: OAuthProvider!
-   
-    
+    private let viewModel = LoginViewModel()
     var body: some View {
         VStack{
             Text("Gitagram")
@@ -20,7 +17,7 @@ struct LoginView: View {
                 .font(.title2)
                 .padding(.bottom,60)
             Button(action: {
-                performOAuthLoginFlow(provider: AuthProvider(rawValue: "GitHub")! )
+                viewModel.performOAuthLoginFlow(provider: AuthProvider(rawValue: "GitHub")! )
             }, label: {
                 HStack{
                     Image("github-mark-white")
@@ -39,46 +36,9 @@ struct LoginView: View {
         }
     }
     
-    private func performOAuthLoginFlow(provider: AuthProvider) {
-      oauthProvider = OAuthProvider(providerID: provider.id)
-       oauthProvider.getCredentialWith(nil) { credential, error in
-         guard error == nil else { 
-             return print(error as Any) }
-         guard let credential = credential else { return }
-         self.signin(with: credential)
-       }
-     }
-    private func signin(with credential: AuthCredential) {
-       Auth.auth().signIn(with: credential) { result, error in
-         guard error == nil else {
-             return print(error as Any)
-         }
-         print("ログインできたよ！")
-       }
-     }
+    
 }
 
 #Preview {
     LoginView()
 }
-
-enum AuthProvider: String {
-    case gitHub = "github.com"
-    var id: String { rawValue }
-
-      /// The UI friendly name of the `AuthProvider`. Used for display.
-    var name: String {
-        switch self {
-        case .gitHub:
-            return "GitHub"
-        }
-    }
-    init?(rawValue: String) {
-        switch rawValue {
-        case "GitHub":
-            self = .gitHub
-        default: return nil
-        }
-    }
-}
-
